+++ changelog.md
@@ -52,11 +52,14 @@ Codion Change Log
 - EntityTablePanel bug fixed, now disables editing for cells based on entities that are neither searchable nor based on small datasets, since a search field can not be automatically instantiated.
 - EntityDialogs.DefaultEditAttributeDialogBuilder refactored, EditAttributePanel added, now handles async properly, validation improved.
 - EditAttributePanel now sets the validation message as component tooltip.
+### is.codion.tools.generator.domain
+- DomainSource now writes a domain unit test.
 ### is.codion.tools.generator.model
 - DomainGeneratorModel.DOMAIN_PACKAGE now with default 'none' instead of 'no.package' since package is a reserved word.
 ### is.codion.tools.generator.panel
 - DomainGeneratorModel, DomainGeneratorPanel, now support separate source paths for api/impl and combined.
 - DomainGeneratorModel, DomainGeneratorPanel, view prefix handling added.
+- DomainGeneratorModel, DomainGeneratorPanel, domain unit test handling added.
 
 ## 0.18.54
 ### is.codion.common.db